parameterize state information
